InternshipTask/requirements.txt
pymupdf
sentence-transformers
faiss-cpu
numpy
requests